create inventory and level system controller and decouple both from the spaceship controller
