+++ Assets/Scripts/InventoryController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47dcd571e098b451ab63e9fb537a5eb1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
